Experiment: use fets to make a remote call to generate npc api
