+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "spa-netlify-ts-monorepo",
+  "name": "traveller-rpg-tools",
   "version": "1.0.0",
   "description": "",
   "type": "module",
